estimate item quantity is also a double

+++ src/main/java/ch/aaap/harvestclient/domain/param/EstimateItemUpdateInfo.java
@@ -22,7 +22,7 @@ public interface EstimateItemUpdateInfo {
     String getDescription();
 
     @Nullable
-    Long getQuantity();
+    Double getQuantity();
 
     @Nullable
     Double getUnitPrice();
